+++ subscription/mongodb/native/blocking/src/test/java/org/occurrent/subscription/mongodb/nativedriver/blocking/NativeMongoSubscriptionModelTest.java
@@ -62,6 +62,7 @@ import static java.time.temporal.ChronoUnit.MILLIS;
 import static java.util.Objects.requireNonNull;
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.catchThrowable;
 import static org.assertj.core.groups.Tuple.tuple;
 import static org.awaitility.Awaitility.await;
 import static org.awaitility.Durations.FIVE_SECONDS;
@@ -116,6 +117,19 @@ public class NativeMongoSubscriptionModelTest {
     }
 
     @Test
+    void blocking_native_mongodb_subscription_throws_iae_when_subscription_already_exists() {
+        // Given
+        String subscriptionId = UUID.randomUUID().toString();
+        subscription.subscribe(subscriptionId, __ -> System.out.println("hello")).waitUntilStarted();
+
+        // When
+        Throwable throwable = catchThrowable(() -> subscription.subscribe(subscriptionId, __ -> System.out.println("hello")).waitUntilStarted());
+        
+        // Then
+        assertThat(throwable).isExactlyInstanceOf(IllegalArgumentException.class).hasMessage("Subscription " + subscriptionId + " is already defined.");
+    }
+
+    @Test
     void blocking_native_mongodb_subscription_calls_listener_for_each_new_event() {
         // Given
         LocalDateTime now = LocalDateTime.now();
@@ -135,7 +149,7 @@ public class NativeMongoSubscriptionModelTest {
     }
 
     @Test
-    void retrying_on_failure() {
+    void  blocking_native_mongodb_subscription_retries_on_failure() {
         // Given
         LocalDateTime now = LocalDateTime.now();
         final AtomicInteger counter = new AtomicInteger(0);