+++ changelog.md
@@ -29,6 +29,7 @@
 * Added retry strategy support to SpringMongoSubscriptionModel. You can define your own by passing an instance of `RetryStrategy` to the constructor. By default
   it'll add a retry strategy with exponential backoff starting with 100 ms and progressively go up to max 2 seconds wait time between 
   each retry when reading/saving/deleting the subscription position.
+* All blocking subscription models will throw an `IllegalArgumentException` if a subscription is registered more than once.
 
 ## Changelog 0.5.1 (2021-01-07)